Refactors List Edit to support Text & Background Colors
